fix: resolve sign comparison warnings in container loops
Replaces int with size_t for vector index variables to eliminate sign comparison warnings reported by clang. This change improves type safety and maintainability while preserving the original functionality of the code. The fix addresses warnings in both part1.cpp and part2.cpp by using consistent unsigned integer types for container operations.

Technical changes:
- Changed loop counter type from int to size_t in all vector iterations
- Modified index variables in both isStrictlySafe and isSafe functions
- Maintains consistent unsigned integer usage for container operations

+++ day02/part1.cpp
@@ -6,7 +6,7 @@ bool isSafe(const vector<int> &levels) {
     return false;
 
   vector<int> diffs;
-  for (int i = 1; i < levels.size(); i++) {
+  for (size_t i = 1; i < levels.size(); i++) {
     diffs.push_back(levels[i] - levels[i - 1]);
   }
